services/payment_service/requirements.txt
Flask
Flask-SQLAlchemy
psycopg2-binary
flask-marshmallow
marshmallow-enum
